**refactor(isobmff): normalize box/length math to `int` and align readers/tests**
- convert ISOBMFF size/remain/length handling from mixed `int64`/`uint*` to `int` across box parsing, metadata traversal, CRX, PRVW, and XPacket paths
- update `XPacketHeader.Length` and related call sites/logging/tests to use `int`
- add explicit bounded conversions (`uint64ToInt`) and keep overflow checks centralized
- refresh ISOBMFF tests for new max-int behavior and remove redundant loop-var copies
- add targeted `nolint:gosec` annotations where casts are spec-bounded

**chore(lint): fix root benchmark/reader lint findings**

- add checked pooled-reader helper in `imagemeta.go` to satisfy `errcheck` type-assertion rules
- replace direct error equality checks with `errors.Is` in benchmarks (`errorlint`)
- remove benchmark shadowed `err` variables (`govet`) and clean pooled-reader error paths

+++ meta/isobmff/log.go
@@ -59,7 +59,7 @@ func logInfoBox(b *box) *zerolog.Event {
 }
 
 func (b *box) log(ev *zerolog.Event) {
-	ev.Str("BoxType", b.boxType.String()).Int64("offset", b.offset).Int64("size", b.size)
+	ev.Str("BoxType", b.boxType.String()).Int64("offset", b.offset).Int("size", b.size)
 	if b.flags != 0 {
 		ev.Object("flags", b.flags)
 	}
@@ -93,7 +93,7 @@ func (c cctpBox) MarshalZerologArray(a *zerolog.Array) {
 
 // MarshalZerologObject is a zerolog interface for logging
 func (b box) MarshalZerologObject(e *zerolog.Event) {
-	e.Str("boxType", b.boxType.String()).Int64("offset", b.offset).Int64("size", b.size)
+	e.Str("boxType", b.boxType.String()).Int64("offset", b.offset).Int("size", b.size)
 	if b.flags != 0 {
 		e.Object("flags", b.flags)
 	}
